requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-genai==1.7.0
youtube-transcript-api==0.6.2
pydantic==2.9.2
